requirements.txt
pdfminer3
pyresparser
streamlit
pandas
pafy
plotly
pymysql
streamlit-tags
Pillow
youtube-dl
nltk
pdfminer3
spacy==3.6.1
time
io
os

# python -m spacy download en_core_web_sm
